🎨 Glyphs as component
glyph animation as component
